crefl: change public api to use decl_ref everywhere
change `crefl_is_*(decl_ref)` to take decl_ref instead of pointers
and concentrate accessor methods that convert between indices,
references and pointers.

the new interface is slightly less efficient codegen-wise, adding
an additional load if inlined, or an additional call when using
extern linkage, but it makes the public interface more consistent
and simpler to use because all interfaces now take `decl_ref`.

this change is being done in preparation for disjunct indices for
builtin types so that the builtin types can be elided from the
binary file representation.

the following interfaces are used to convert from indices to or
from references, and to access decl properties:

- decl * crefl_ptr(decl_ref d);
- decl_tag crefl_tag(decl_ref d);
- decl_set crefl_attrs(decl_ref d);
- decl_id crefl_idx(decl_ref d);
- decl_ref crefl_next(decl_ref d);
- decl_ref crefl_lookup(decl_db *db, size_t decl_idx);

these methods are currently default to extern. the internal code
could be made more efficient by using extern inline, so that the
internal routines inline the accessors, however, extern inline is
not supported by C++, only by C.

after this change, the only other methods that assume the internal
structure of decl_db is a linear indexed array are the following:

- crefl_new
- crefl_db_new
- crefl_read_db
- crefl_write_db
- crefl_find_intrinsic
- _decl_top_level_fetch

the intention is a future change will start builtin types at 0
while user-defined types start at some other offset e.g. 1024.