crefl: change crefl_constant_value return type to decl_raw
variable, uniform and function return values need to be
determined at final link because they are dependent on final
link, and can be of a variable width, however constant values
can be determined at compile-time and on current architectures
their value is limited to 64-bit hence they can return decl_raw.

these values can be determined at compile-time:

  - decl_raw crefl_constant_value(decl_ref d);

these values can be determined at link-time:

  - void * crefl_variable_addr(decl_ref d);
  - void * crefl_uniform_addr(decl_ref d);
  - void * crefl_function_addr(decl_ref d);

object address metadata is not yet implemented and will require a
module to be emitted and compiled to reference the objects and
take their addresses in a per assembly array of structs indexed
by some link-time reflection identity such as a hash of the fully
qualified object name.

it is currently believed that decl_raw, which is 64-bit, should
suffice for constant values. ideally one could use a bignum for
wider values supported by the abstract machine, although that will
not fit within the current node space and would need a id and size
for a value in the symbol space.