cfileio: change the file magic underscore into a zero
decided to not waste the underscore and instead recruit it as an
extra digit for the version number. the version number is hence
changed to '000'. this is work-in-progress alpha code and the file
format is subject to change. the first stable version will likely
increment the counter and use '001', although it could be that the
magic changes and a version field is added to the header struct.