crefl: crefl_db_read_mem must use realloc not malloc
this would be a relatively major security bug for production code,
although the code is still prerelease and undergoing testing, and
this was undercovered during construction of test cases.

the intention of the code is to bulk copy array members, and the
bug means the symbol table entries and strings become undefined if
the initial size is not sufficient for the records in the buffer.