model: simplify decl struct by moving _link and adding quantity
- move _link field from union into decl base as the link field
  is common to almost subtypes.
- rename the array 'size' property to 'count' because it is less
  ambiguous. size could refer to a concrete size of the array.
- rename crefl_array_size to crefl_array_count

after moving the _decl link field into the base were left with
some quantities in the union elements. i.e.

  struct { decl_sz _width; } _decl_intrinsic;
  struct { decl_sz _width; } _decl_field;
  struct { decl_sz _count; } _decl_array;
  struct { decl_sz _value; } _decl_constant;
  struct { decl_sz _addr;  } _decl_function;

simpler to just provide a universal type specific quantity. i.e.

  union {
      decl_sz _quantity;
      decl_sz _width;
      decl_sz _count;
      decl_sz _value;
      decl_sz _addr;
  };

this is less constrained, but we have no mechanism in C to tie
a tagged union type tag to the union types, so we decided to
instead focus on source readibility and simplicity, as this
makes the code more compact.