crefl: make list functions ignore input count for null array
the crefl_list_* functions uses the pointer to array and pointer
to count pattern, and were meant to ignore array writes if the
array is null, however, the implemented semantic was to ignore
the array writes if the count was zero.

modify t3/adjacent-structs-1.h test to fetch with a non-zero
count and a null array to make sure this combination works.