cmodel: rename _attrs to _props in preparation for C attributes
we can't occupy attributes for type properties if we want to
add support for C attributes because there will be a name clash.

refer to type specific properties as _props not _attrs. we will
later add decl_id _attr as an additional link so that any node
type can support attributes.