crefl: add in and out qualifiers and label return parameter out
function return value and parameters are currently stored in a single
param list refered to in the link field of the function element. since
there is only a single link field, the return parameter is simply the
first element in the list, and it is anonymous, because C return params
are unnamed.

 (A a, B b, C c) -> (E e, F f) ~=  (out E, out F, A a, B b, C c)

to make the return parameter more clear, it is now labelled with 'out'.
'in' and 'out' are otherwise currently not used.

e.g.

===== INPUT  test-case: test/cv-qualifier-2.h =====

_Complex double h(void * restrict a, const int b);

===== OUTPUT test-case: test/cv-qualifier-2.h =====

id    attr  next  link  type       name         props        link-detail
----- ----- ----- ----- ---------- ------------ ------------ ------------------
23    0     0     24    function   h            addr=0x0     param(anonymous)
24    0     25    21    param      (anonymous)  out          intrinsic("cdouble")
25    0     27    26    param      a            restrict     pointer("*")
26    0     0     0     pointer    *            width=64     void(anonymous)
27    0     0     10    param      b            const        intrinsic("int")
----- ----- ----- ----- ---------- ------------ ------------ ------------------