crefl: make sure void* pointee type is intrinsic void
change get_intrinsic_type fallthrough to select void if no other type
is matched. _void property does not have a value but the check in
get_intrinsic_type will match due to the form of the properties test,
which works on the basis of finding an intrinsic with width 0.

see this clause from crefl_intrinsic:

        if (crefl_decl_tag(d) == _decl_intrinsic &&
            crefl_decl_ptr(d)->_width == width &&
                ((crefl_decl_props(d) & props) == props)) {

so we change get_intrinsic type final else to find void:

            tr = crefl_intrinsic(db, _void, t.Width);

this means when the pointer is peeled from pointer to void we will link
to the void type (1) intrinsic("void") instead of the empty type entry
(0) none(anonymous). see diff:

===== INPUT  test-case: test/cv-qualifier-2.h =====

_Complex double h(void * restrict a, const int b);

===== OUTPUT test-case: test/cv-qualifier-2.h =====

 ----- ----- ----- ----- ---------- ------------ ------------ ------------------
 23    0     0     24    function   h            addr=0x0     param(anonymous)
 24    0     25    21    param      (anonymous)  out          intrinsic("cdouble")
-25    0     27    26    param      a            restrict     pointer("*")
-26    0     0     0     pointer    *            width=64     none(anonymous)
+25    0     27    26    param      a            restrict     pointer("void*")
+26    0     0     1     pointer    void*        width=64     intrinsic("void")
 27    0     0     10    param      b            const        intrinsic("int")
 ----- ----- ----- ----- ---------- ------------ ------------ ------------------