crefl: rename empty type system tag from void to none
void is a misnomer for the internal none type because C void is a
subtype property of intrinsic. The use of 'none' for the empty tag is
more accurate because it is used in the 'link' field of struct or
union types to indicate an empty struct or union and it is used in
the 'next' field to signify the end of lists, and in 'attr' field to
indicate the absence of attributes. Thus we rename it to 'none'.

e.g.

===== INPUT  test-case: test/empty-struct-1.h =====

struct a {};

===== OUTPUT test-case: test/empty-struct-1.h =====

id    attr  next  link  type       name         props        link-detail
----- ----- ----- ----- ---------- ------------ ------------ ------------------
23    0     0     0     struct     a                         none(anonymous)
----- ----- ----- ----- ---------- ------------ ------------ ------------------