creflect: fix parent mismatch by comparing stack to current parent
the visitor interface does not provide any depth information
so pops need to be detected. the code this patch replaces did
not detect all cases, in particular if the last element in a
sequence points to a next decl node that is not visited.

make pop detection robust by comparing parent with the parent
id on the stack, and loop popping until they match. this method
works even when the last node before ascending was not visited.
the new method is simpler and less brittle than the old method.