cmake: add crefl_target_reflect macro to link reflection metadata
the crefl_target_reflect macro adds custom commands that reflect
on the sources for a target using creflcc, then merge the resulting
reflection metadata into an archive, and compile it into a library.

the reflection metadata archive library can then be linked with
the exectuable to allow use of crefl_db_read_mem. e.g.

    decl_db *db = crefl_db_new();
    crefl_db_read_mem(db, __crefl_main_data, __crefl_main_size);