cmake: connect up crefl_target_reflect macro to example2_embed
this closes the loop on linking and deduplication of metadata,
which is emitted to a library and linked with the executable.

- reverse crefl_target_reflect operands so that target is first.
- use full relative path in reflection output files so that
  files with the same name in different directories don't clash.
- rename example1 file api to example1_file/main.c
- add embed example named example2_embed/main.c

example2_embed uses the following pattern to read the reflection
metadata into memory. this will ultimately be wrapped with macros
that stringify a type name and return the metadata for a type.

--

extern const char __crefl_main_data[];
extern const size_t __crefl_main_size;

int main(int argc, const char **argv)
{
    decl_db *db = crefl_db_new();
    crefl_db_read_mem(db, __crefl_main_data, __crefl_main_size);

    ...

    crefl_db_destroy(db);
}